structure files into category
