+++ data_structures/Binary Tree Traversal/preorder.JPG
