+++ geometric_algo/triangle.c
