+++ number_theory/armstrong.c
