+++ number_theory/conversions/hex_bin.c
