+++ number_theory/conversions/hex_dec.c
