+++ number_theory/palindrome.c
