+++ number_theory/pow.c
