+++ number_theory/squrt.c
