+++ patterns/pattern_1.c
