+++ search/readme.md
@@ -1,0 +1,27 @@
+### Search
+
+1. Linear Search
+2. Binary Search
+3. Jump Search
+4. Interpolation Search
+5. Exponential Search
+
+
+### Linear Search
+
+Linear Search is a search algorithm that sequentially checks each element in the list to see if it is the one we are looking for. The time complexity of Linear Search is O(n).
+### Binary Search
+
+Binary search is one of the most efficient way to search a sorted array. It works by repeatedly dividing half the portion of the list that could 
+contain the target item, until you've narrowed down the possible locations to just one.
+
+
+a very simple explanation of this algorithm can be found here 
+
+CS50 2020 https://youtu.be/gR6nycuZKlM?t=1378
+
+Time complexity of this algorithm is O(logn) (read as O of logn), that means in an array of 12 elements in the worst case you will have to make 3 look ups. while incase of a linear search it would be 11 look ups. 
+
+
+
+
