+++ series/factorial.c
