+++ series/sumprime.c
