+++ string_manipulation/camel.c
