+++ string_manipulation/camel_cmd.c
