+++ string_manipulation/eliminatevowels.c
