+++ string_manipulation/string_palindrome.c
